rag-vs-longcontext/requirements.txt
datasets
faiss-cpu==1.8.0
sentence-transformers
openai>=1.0.0
numpy
tqdm
PyYAML
regex
